Basic Scaffolding of media module
Added a readme file for the module based on the README Template: https://www.drupal.org/node/2181737.
Added a module file with a simple help function which provides help for the module.
Added an info file which lists the dependencies and basic module info
Added a .travis.yml file for TRAVIS CI
Added a composer.json file which defined the project.